added code to display the bill of materials. there is a bug when displaying a flat roof carport though..

+++ src/main/java/Data/Entity/Carport.java
@@ -28,7 +28,7 @@ public class Carport {
         return width;
     }
 
-    public double getInclination() {
+    public int getInclination() {
         return inclination;
     }
     public int getLength() {
